chore: extend incremental filename enforcement to api-client user
