+++ libraries/api-client/src/conversation/RemoteConversations.ts
@@ -19,7 +19,7 @@
 
 import {Conversation} from './Conversation';
 
-import {QualifiedId} from '../user/QualifiedId';
+import {QualifiedId} from '../user/qualifiedId';
 
 export interface RemoteConversations {
   failed?: QualifiedId[];
