+++ libraries/api-client/src/user/completePasswordReset.ts
