+++ libraries/api-client/src/user/data/userDeleteData.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2021 Wire Swiss GmbH
+ * Copyright (C) 2019 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,8 +17,9 @@
  *
  */
 
-import {QualifiedId} from './QualifiedId';
+import {QualifiedId} from '../qualifiedId';
 
-export interface LimitedQualifiedUserIdList {
-  qualified_users: QualifiedId[];
+export interface UserDeleteData {
+  id: string;
+  qualified_id: QualifiedId;
 }
