+++ libraries/api-client/src/user/data/userPropertiesDeleteData.ts
