+++ libraries/api-client/src/user/index.ts
@@ -17,25 +17,26 @@
  *
  */
 
-export * from './Activate';
-export * from './ActivationResponse';
-export * from './CheckHandles';
-export * from './CompletePasswordReset';
-export * from './Contact';
-export * from './HandleInfo';
-export * from './LimitedQualifiedUserIdList';
-export * from './ManagedSource';
-export * from './NewPasswordReset';
-export * from './QualifiedHandle';
-export * from './QualifiedId';
-export * from './RichInfo';
-export * from './RTCIceServer';
-export * from './SearchResult';
-export * from './SendActivationCode';
-export * from './User';
-export * from './UserAPI';
-export * from './UserAsset';
-export * from './UserError';
-export * from './UserPreKeyBundleMap';
-export * from './UserUpdate';
-export * from './VerifyDelete';
+export * from './activate';
+export * from './activationResponse';
+export * from './checkHandles';
+export * from './completePasswordReset';
+export * from './contact';
+export * from './handleInfo';
+export * from './limitedQualifiedUserIdList';
+export * from './managedSource';
+export * from './newPasswordReset';
+export * from './qualifiedHandle';
+export * from './qualifiedId';
+export * from './richInfo';
+export * from './rtcConfiguration';
+export * from './rtcIceServer';
+export * from './searchResult';
+export * from './sendActivationCode';
+export * from './user';
+export * from './userApi';
+export * from './userAsset';
+export * from './userError';
+export * from './userPreKeyBundleMap';
+export * from './userUpdate';
+export * from './verifyDelete';
