+++ libraries/api-client/src/user/richInfo.ts
