+++ libraries/api-client/src/user/searchResult.ts
