+++ libraries/api-client/src/user/sendActivationCode.ts
