+++ libraries/api-client/src/user/verifyDelete.ts
