+++ libraries/api-client/src/userGroups/userGroup.ts
@@ -19,7 +19,7 @@
 
 import {UserGroupManagedBy} from './userGroupManagedBy';
 
-import {QualifiedId} from '../user/QualifiedId';
+import {QualifiedId} from '../user/qualifiedId';
 
 export interface UserGroup {
   createdAt: string;
